requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
psycopg==3.2.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
